+++ shardis-ui/src/main/frontend/platform/browser/providers.ts
@@ -6,7 +6,9 @@ import {FORM_PROVIDERS} from 'angular2/common';
 // Angular 2 Http
 import {HTTP_PROVIDERS} from 'angular2/http';
 // Angular 2 Router
-import {ROUTER_PROVIDERS, LocationStrategy, PathLocationStrategy} from 'angular2/router';
+import {ROUTER_PROVIDERS} from 'angular2/router';
+import {LocationStrategy} from 'angular2/src/platform/browser/location/location_strategy';
+import {PathLocationStrategy} from 'angular2/src/platform/browser/location/path_location_strategy';
 
 // Angular 2 Material
 // import {MdRadioDispatcher} from '@angular2-material/radio/radio_dispatcher';