+++ shardis-ui/src/main/frontend/app/views/playground/heroes/app.component.ts
@@ -9,7 +9,7 @@ import {HeroService} from './hero.service.ts';
     <h1>{{title}}</h1>
     <h2>My Heroes</h2>
     <ul class="heroes">
-      <li *ngFor="#hero of heroes"
+      <li *ngFor="let hero of heroes"
         [class.selected]="hero === selectedHero"
         (click)="onSelect(hero)">
         <span class="badge">{{hero.id}}</span> {{hero.name}}