+++ shardis-ui/src/main/frontend/app/components/foundation/accordion/accordion.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit, Input} from 'angular2/core';
+import {Component, OnInit, Input} from '@angular/core';
 import {AccordionTab} from './accordion.tab.component';
 
 @Component({