+++ shardis-ui/src/main/frontend/app/directives/x-large.directive.spec.ts
@@ -4,12 +4,12 @@ import {
   async,
   describe,
   beforeEachProviders,
-  TestComponentBuilder
-} from 'angular2/testing';
+} from '@angular/core/testing';
 
-import {Component, provide} from 'angular2/core';
-import {BaseRequestOptions, Http} from 'angular2/http';
-import {MockBackend} from 'angular2/http/testing';
+import {Component, provide} from '@angular/core';
+import {BaseRequestOptions, Http} from '@angular/http';
+import {MockBackend} from '@angular/http/testing';
+import {TestComponentBuilder} from '@angular/compiler/testing';
 
 // Load the implementations that should be tested
 import {XLarge} from './x-large.directive.ts';