+++ shardis-ui/src/main/frontend/app/views/about/about.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit} from 'angular2/core';
+import {Component, OnInit} from '@angular/core';
 
 console.log('`About` component loaded asynchronously');