+++ shardis-ui/src/main/frontend/app/views/home/services/title.service.spec.ts
@@ -2,18 +2,17 @@ import {
   it,
   inject,
   async,
-  beforeEachProviders,
-  TestComponentBuilder
-} from 'angular2/testing';
+  beforeEachProviders
+} from '@angular/core/testing';
 
-import {Component, provide} from 'angular2/core';
-import {BaseRequestOptions, Http} from 'angular2/http';
-import {MockBackend} from 'angular2/http/testing';
+import {Component, provide} from '@angular/core';
+import {BaseRequestOptions, Http} from '@angular/http';
+import {MockBackend} from '@angular/http/testing';
 
 
 import {Title} from './title.service.ts';
-import {ResponseOptions} from 'angular2/http';
-import {Response} from 'angular2/http';
+import {ResponseOptions} from '@angular/http';
+import {Response} from '@angular/http';
 import {AuthService} from '../../../services/auth.service';
 
 describe('Title', () => {