+++ shardis-ui/src/main/frontend/app/views/kitchensink/kitchensink.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit} from 'angular2/core';
+import {Component, OnInit} from '@angular/core';
 
 
 console.log('`KitchenSink` component loaded asynchronously');