+++ shardis-ui/src/main/frontend/platform/browser/router-active/index.ts
@@ -1,2 +1,2 @@
 // Application level directive
-export * from './router-active';
+export * from './router-active.directive';