+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Changelog
 
 ### 1.4.0 (UNRELEASED)
+* updated Angular to beta 16
+* updated node dependencies
+* updated maven dependencies
 
 ### 1.3.0 (21.07.2016)
 * refactoring to microservices