+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Changelog
 
-### 1.4.0 (UNRELEASED)
+### 1.4.0 (05.05.2016)
 * Ppdated Angular to beta 16
 * Updated node dependencies
 * Updated maven dependencies